[Mutable] Move compilation-related data to the FModelResource struct. Change type for NumLODs, FirstLODAvailable, etc.
#jira UE-205600
#rnx
#rb gerard.martin

[CL 32400962 by pere rifa in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/CustomizableObject/Internal/MuCO/CustomizableObjectSystemPrivate.h
@@ -513,9 +513,9 @@ public:
 	mu::Instance::ID InstanceID = 0; // Redundant
 	const mu::Instance* MutableInstance = nullptr;
 
-	int32 NumComponents = 0;
-	int32 NumLODsAvailable = 0;
-	int32 FirstLODAvailable = 0;
+	uint8 NumComponents = 0;
+	uint8 NumLODsAvailable = 0;
+	uint8 FirstLODAvailable = 0;
 	uint8 FirstResidentLOD = 0;
 
 	TMap<uint32, FTexturePlatformData*> ImageToPlatformDataMap;
